refactor: move number related functions
